+++ webknossos/webknossos/dataset/dataset.py
@@ -35,6 +35,8 @@ from natsort import natsort_keygen
 from numpy.typing import DTypeLike
 from upath import UPath
 
+from webknossos.geometry.vec_int import VecInt, VecIntLike
+
 from ..client.api_client.models import ApiDataset
 from ..geometry.vec3_int import Vec3Int, Vec3IntLike
 from ._array import ArrayException, ArrayInfo, BaseArray
@@ -60,7 +62,7 @@ if TYPE_CHECKING:
     from ..administration.user import Team
     from ..client._upload_dataset import LayerToLink
 
-from ..geometry import BoundingBox, Mag
+from ..geometry import BoundingBox, Mag, NDBoundingBox
 from ..utils import (
     copy_directory_with_symlinks,
     copytree,
@@ -783,7 +785,7 @@ class Dataset:
         dtype_per_channel: Optional[DTypeLike] = None,
         num_channels: Optional[int] = None,
         data_format: Union[str, DataFormat] = DEFAULT_DATA_FORMAT,
-        bounding_box: Optional[BoundingBox] = None,
+        bounding_box: Optional[NDBoundingBox] = None,
         **kwargs: Any,
     ) -> Layer:
         """
@@ -1048,7 +1050,7 @@ class Dataset:
         compress: bool = False,
         *,
         ## other arguments
-        topleft: Vec3IntLike = Vec3Int.zeros(),  # in Mag(1)
+        topleft: VecIntLike = Vec3Int.zeros(),  # in Mag(1)
         swap_xy: bool = False,
         flip_x: bool = False,
         flip_y: bool = False,
@@ -1215,8 +1217,12 @@ class Dataset:
                 num_channels=pims_images.num_channels,
                 **add_layer_kwargs,  # type: ignore[arg-type]
             )
+
+            expected_bbox = pims_images.expected_bbox
+
+            # When the expected bbox is 2D the chunk_shape is set to 2D too.
             if (
-                pims_images.expected_shape.z == 1
+                expected_bbox.get_shape("z") == 1
                 and layer.data_format == DataFormat.Zarr
             ):
                 if chunk_shape is None:
@@ -1227,17 +1233,13 @@ class Dataset:
             if chunks_per_shard is None and layer.data_format == DataFormat.Zarr3:
                 chunks_per_shard = DEFAULT_CHUNKS_PER_SHARD_FROM_IMAGES
 
+            mag = Mag(mag)
+            layer.bounding_box = expected_bbox.from_mag_to_mag1(mag).offset(topleft)
             mag_view = layer.add_mag(
                 mag=mag,
                 chunk_shape=chunk_shape,
                 chunks_per_shard=chunks_per_shard,
                 compress=compress,
-            )
-            mag = mag_view.mag
-            layer.bounding_box = (
-                BoundingBox((0, 0, 0), pims_images.expected_shape)
-                .from_mag_to_mag1(mag)
-                .offset(topleft)
             )
 
             if batch_size is None:
@@ -1262,10 +1264,44 @@ class Dataset:
             )
 
             args = []
-            for z_start in range(0, pims_images.expected_shape.z, batch_size):
-                z_end = min(z_start + batch_size, pims_images.expected_shape.z)
-                # return shapes and set to union when using --pad
-                args.append((z_start, z_end))
+            bbox = layer.bounding_box
+            additional_axes = [
+                axis_name for axis_name in bbox.axes if axis_name not in ("x", "y", "z")
+            ]
+            additional_axes_shapes = tuple(
+                product(
+                    *[range(bbox.get_shape(axis_name)) for axis_name in additional_axes]
+                )
+            )
+            if additional_axes and layer.data_format != DataFormat.Zarr3:
+                assert (
+                    len(additional_axes_shapes) == 1
+                ), "The data stores additional axes with shape bigger than 1. These are only supported by data format Zarr3."
+
+                # Convert NDBoundingBox to 3D BoundingBox
+                bbox = BoundingBox(
+                    bbox.topleft_xyz,
+                    bbox.size_xyz,
+                )
+                expected_bbox = bbox
+                additional_axes = []
+
+            z_shape = bbox.get_shape("z")
+            bbox = bbox.with_topleft(VecInt.zeros(bbox.axes))
+            for z_start in range(0, z_shape, batch_size):
+                z_size = min(batch_size, z_shape - z_start)
+                z_bbox = bbox.with_bounds("z", z_start, z_size)
+                if not additional_axes:
+                    args.append(z_bbox)
+                else:
+                    for shape in additional_axes_shapes:
+                        reduced_bbox = z_bbox
+                        for index, axis in enumerate(additional_axes):
+                            reduced_bbox = reduced_bbox.with_bounds(
+                                axis, shape[index], 1
+                            )
+                        args.append(reduced_bbox)
+
             with warnings.catch_warnings():
                 # Block alignmnent within the dataset should not be a problem, since shard-wise chunking is enforced.
                 # However, dataset borders might change between different parallelized writes, when sizes differ.
@@ -1294,18 +1330,14 @@ class Dataset:
                 if category == "segmentation":
                     max_id = max(max_ids)
                     cast(SegmentationLayer, layer).largest_segment_id = max_id
-                actual_size = Vec3Int(
-                    dimwise_max(shapes) + (pims_images.expected_shape.z,)
+                layer.bounding_box = layer.bounding_box.with_size_xyz(
+                    Vec3Int(dimwise_max(shapes) + (layer.bounding_box.get_shape("z"),))
+                    * mag.to_vec3_int().with_z(1)
                 )
-                layer.bounding_box = (
-                    BoundingBox((0, 0, 0), actual_size)
-                    .from_mag_to_mag1(mag)
-                    .offset(topleft)
-                )
-            if pims_images.expected_shape != actual_size:
+            if expected_bbox != layer.bounding_box:
                 warnings.warn(
                     "[WARNING] Some images are larger than expected, smaller slices are padded with zeros now. "
-                    + f"New size is {actual_size}, expected {pims_images.expected_shape}."
+                    + f"New bbox is {layer.bounding_box}, expected {expected_bbox}."
                 )
             if first_layer is None:
                 first_layer = layer
@@ -1515,7 +1547,7 @@ class Dataset:
         self._export_as_json()
         return self.layers[new_layer_name]
 
-    def calculate_bounding_box(self) -> BoundingBox:
+    def calculate_bounding_box(self) -> NDBoundingBox:
         """
         Calculates and returns the enclosing bounding box of all data layers of the dataset.
         """
@@ -1736,12 +1768,19 @@ class Dataset:
         self._ensure_writable()
 
         properties_on_disk = self._load_properties()
-
-        if properties_on_disk != self._last_read_properties:
+        try:
+            if properties_on_disk != self._last_read_properties:
+                warnings.warn(
+                    "[WARNING] While exporting the dataset's properties, properties were found on disk which are "
+                    + "newer than the ones that were seen last time. The properties will be overwritten. This is "
+                    + "likely happening because multiple processes changed the metadata of this dataset."
+                )
+        except ValueError:
+            # the __eq__ operator raises a ValueError when two bboxes are not comparable. This is the case when the
+            # axes are not the same. During initialization axes are added or moved sometimes.
             warnings.warn(
-                "[WARNING] While exporting the dataset's properties, properties were found on disk which are "
-                + "newer than the ones that were seen last time. The properties will be overwritten. This is "
-                + "likely happening because multiple processes changed the metadata of this dataset."
+                "[WARNING] Properties changed in a way that they are not comparable anymore. Most likely "
+                + "the bounding box naming or axis order changed."
             )
 
         with (self.path / PROPERTIES_FILE_NAME).open("w", encoding="utf-8") as outfile: